:sparkles: Feat: 아이템 상세 조회 모달창 적용
