+++ src/components/ContestListButton/styled.js
@@ -1,5 +1,18 @@
 import styled from 'styled-components';
 
+/**
+ * 콘테스트 목록가기 버튼 styled components
+ * @author 정은찬
+ * @since 2024.09.12
+ * @version 1.0
+ *
+ * <pre>
+ * 수정일       수정자        수정내용
+ * ----------  --------    ---------------------------
+ * 2024.09.12  	정은찬        최초 생성
+ * </pre>
+ */
+
 export const StyledButton = styled.button`
   border-radius: 50px;
   background-color: #284180;
