+++ src/components/CoordinationPost/index.jsx
@@ -16,6 +16,19 @@ import HeartEmpty from '../../assets/images/heart-empty-icon.svg';
 import HeartFull from '../../assets/images/heart-full-icon.svg';
 import Text from '../../components/Text';
 
+/**
+ * 코디 카드 컴포넌트
+ * @author 정은찬
+ * @since 2024.09.13
+ * @version 1.0
+ *
+ * <pre>
+ * 수정일       수정자        수정내용
+ * ----------  --------    ---------------------------
+ * 2024.09.13  	정은찬        최초 생성
+ * </pre>
+ */
+
 const CoordinationPost = ({ post }) => {
   const navigate = useNavigate(); // useNavigate 훅 사용
