+++ src/components/RecommendButton/styled.js
@@ -1,5 +1,18 @@
 import styled from 'styled-components';
 
+/**
+ * 추천하기 버튼 styled component
+ * @author 정은찬
+ * @since 2024.09.13
+ * @version 1.0
+ *
+ * <pre>
+ * 수정일       수정자        수정내용
+ * ----------  --------    ---------------------------
+ * 2024.09.13  	정은찬        최초 생성
+ * </pre>
+ */
+
 export const StyledButton = styled.button`
   border-radius: 50px;
   background-color: #284180;
